refactor: project layer architecture (presentation, business, persistence)
