+++ src/main/java/com/example/realworld/application/articles/dto/ResponseArticle.java
@@ -2,8 +2,8 @@ package com.example.realworld.application.articles.dto;
 
 import com.example.realworld.application.articles.persistence.Article;
 import com.example.realworld.application.tags.persistence.Tag;
-import com.example.realworld.application.users.persistence.User;
 import com.example.realworld.application.users.dto.ResponseProfile;
+import com.example.realworld.application.users.persistence.User;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AccessLevel;
 import lombok.Getter;
