+++ src/main/java/com/example/realworld/application/articles/persistence/Article.java
@@ -4,7 +4,7 @@ import com.example.realworld.application.articles.exception.NotFoundCommentExcep
 import com.example.realworld.application.favorites.persistence.FavoriteArticle;
 import com.example.realworld.application.tags.persistence.Tag;
 import com.example.realworld.application.users.persistence.User;
-import com.example.realworld.core.domain.BaseTimeEntity;
+import com.example.realworld.core.persistence.BaseTimeEntity;
 import lombok.*;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.util.StringUtils;
