+++ src/main/java/com/example/realworld/application/favorites/business/FavoriteArticleBusinessService.java
@@ -1,11 +1,11 @@
 package com.example.realworld.application.favorites.business;
 
-import com.example.realworld.application.articles.persistence.Article;
-import com.example.realworld.application.articles.persistence.ArticleDomainService;
 import com.example.realworld.application.articles.dto.ResponseArticle;
 import com.example.realworld.application.articles.exception.DuplicateFavoriteArticleException;
-import com.example.realworld.application.favorites.persistence.FavoriteArticle;
+import com.example.realworld.application.articles.persistence.Article;
+import com.example.realworld.application.articles.persistence.ArticleDomainService;
 import com.example.realworld.application.favorites.exception.NotYetFavoriteArticleException;
+import com.example.realworld.application.favorites.persistence.FavoriteArticle;
 import com.example.realworld.application.favorites.persistence.repository.FavoriteArticleRepository;
 import com.example.realworld.application.users.persistence.User;
 import com.example.realworld.application.users.persistence.UserDomainService;
