+++ src/main/java/com/example/realworld/application/favorites/persistence/FavoriteArticle.java
@@ -2,7 +2,7 @@ package com.example.realworld.application.favorites.persistence;
 
 import com.example.realworld.application.articles.persistence.Article;
 import com.example.realworld.application.users.persistence.User;
-import com.example.realworld.core.domain.BaseTimeEntity;
+import com.example.realworld.core.persistence.BaseTimeEntity;
 import lombok.*;
 
 import javax.persistence.*;
