+++ src/main/java/com/example/realworld/application/follows/business/FollowBusinessService.java
@@ -1,12 +1,12 @@
 package com.example.realworld.application.follows.business;
 
-import com.example.realworld.application.follows.persistence.Follow;
 import com.example.realworld.application.follows.exception.DuplicateFollowException;
 import com.example.realworld.application.follows.exception.NotFoundFollowException;
+import com.example.realworld.application.follows.persistence.Follow;
 import com.example.realworld.application.follows.persistence.repository.FollowRepository;
+import com.example.realworld.application.users.dto.ResponseProfile;
 import com.example.realworld.application.users.persistence.User;
 import com.example.realworld.application.users.persistence.UserDomainService;
-import com.example.realworld.application.users.dto.ResponseProfile;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
