+++ src/test/java/com/example/realworld/application/articles/business/ArticleBusinessServiceTest.java
@@ -1,8 +1,8 @@
 package com.example.realworld.application.articles.business;
 
-import com.example.realworld.application.articles.persistence.Article;
 import com.example.realworld.application.articles.dto.*;
 import com.example.realworld.application.articles.exception.NotFoundArticleException;
+import com.example.realworld.application.articles.persistence.Article;
 import com.example.realworld.application.articles.persistence.repository.ArticleRepository;
 import com.example.realworld.application.users.persistence.User;
 import com.example.realworld.application.users.persistence.repository.UserRepository;
