+++ src/test/java/com/example/realworld/application/users/repository/UserRepositoryTest.java
@@ -1,7 +1,7 @@
 package com.example.realworld.application.users.repository;
 
-import com.example.realworld.application.users.persistence.User;
 import com.example.realworld.application.users.exception.NotFoundUserException;
+import com.example.realworld.application.users.persistence.User;
 import com.example.realworld.application.users.persistence.repository.UserRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
